chore: store cidr in runtimeconfig
store cidr in the runtimeconfig.
