+++ cmd/embedded-cluster/install.go
@@ -409,7 +409,7 @@ func ensureK0sConfig(c *cli.Context, provider *defaults.Provider, applier *addon
 	cfg.Spec.API.Address = address
 	cfg.Spec.Storage.Etcd.PeerAddress = address
 
-	podCIDR, serviceCIDR, err := DeterminePodAndServiceCIDRs(c)
+	podCIDR, serviceCIDR, err := DeterminePodAndServiceCIDRs(c, provider)
 	if err != nil {
 		return nil, fmt.Errorf("unable to determine pod and service CIDRs: %w", err)
 	}
@@ -692,6 +692,7 @@ func installCommand() *cli.Command {
 				getDataDirFlag(runtimeConfig),
 				getAdminConsolePortFlag(runtimeConfig),
 				getLocalArtifactMirrorPortFlag(runtimeConfig),
+				getNetworkCIDRFlag(runtimeConfig),
 			},
 		)),
 		Action: func(c *cli.Context) error {
@@ -701,12 +702,12 @@ func installCommand() *cli.Command {
 			defer tryRemoveTmpDirContents(provider)
 
 			var err error
-			proxy, err := getProxySpecFromFlags(c)
+			proxy, err := getProxySpecFromFlags(c, provider)
 			if err != nil {
 				return fmt.Errorf("unable to get proxy spec from flags: %w", err)
 			}
 
-			proxy, err = includeLocalIPInNoProxy(c, proxy)
+			proxy, err = includeLocalIPInNoProxy(c, proxy, provider)
 			if err != nil {
 				metrics.ReportApplyFinished(c, err)
 				return err
