+++ cmd/embedded-cluster/proxy_test.go
@@ -6,6 +6,7 @@ import (
 
 	k0sconfig "github.com/k0sproject/k0s/pkg/apis/k0s/v1beta1"
 	ecv1beta1 "github.com/replicatedhq/embedded-cluster/kinds/apis/v1beta1"
+	"github.com/replicatedhq/embedded-cluster/pkg/defaults"
 	"github.com/stretchr/testify/assert"
 	"github.com/urfave/cli/v2"
 )
@@ -109,26 +110,15 @@ func Test_getProxySpecFromFlags(t *testing.T) {
 				NoProxy:         "localhost,127.0.0.1,.cluster.local,.svc,other-no-proxy-1,other-no-proxy-2,1.1.1.1/24,2.2.2.2/24",
 			},
 		},
-		{
-			name: "--cidr should override default no proxy",
-			init: func(t *testing.T, flagSet *flag.FlagSet) {
-				flagSet.Set("http-proxy", "http://other-proxy")
-				flagSet.Set("https-proxy", "https://other-proxy")
-				flagSet.Set("no-proxy", "other-no-proxy-1,other-no-proxy-2")
-
-				flagSet.Set("cidr", "10.0.0.0/16")
-			},
-			want: &ecv1beta1.ProxySpec{
-				HTTPProxy:       "http://other-proxy",
-				HTTPSProxy:      "https://other-proxy",
-				ProvidedNoProxy: "other-no-proxy-1,other-no-proxy-2",
-				NoProxy:         "localhost,127.0.0.1,.cluster.local,.svc,other-no-proxy-1,other-no-proxy-2,10.0.0.0/17,10.0.128.0/17",
-			},
-		},
 	}
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			flags := withProxyFlags(withSubnetCIDRFlags([]cli.Flag{}))
+			runtimeConfig := &ecv1beta1.RuntimeConfigSpec{}
+			provider := defaults.NewProviderFromRuntimeConfig(runtimeConfig)
+
+			flags := []cli.Flag{getNetworkCIDRFlag(runtimeConfig)}
+			flags = withProxyFlags(withSubnetCIDRFlags(flags))
+
 			flagSet := flag.NewFlagSet("test", 0)
 			for _, flag := range flags {
 				flag.Apply(flagSet)
@@ -136,8 +126,9 @@ func Test_getProxySpecFromFlags(t *testing.T) {
 			if tt.init != nil {
 				tt.init(t, flagSet)
 			}
+
 			c := cli.NewContext(cli.NewApp(), flagSet, nil)
-			got, err := getProxySpecFromFlags(c)
+			got, err := getProxySpecFromFlags(c, provider)
 			assert.NoError(t, err, "unexpected error received")
 			assert.Equal(t, tt.want, got)
 		})
